Update depend-inj.md
Proxies can also be used to break cyclical dependencies (where class A depends on class B and class B depends on class A). There's a specific exception that is thrown when the ObjectManager detects a cycle ( https://github.com/magento/magento2/blob/master/lib/internal/Magento/Framework/ObjectManager/Factory/Dynamic/Developer.php#L80 ) in non-production mode. One way to resolve this is to use a Proxy as a constructor argument in the di.xml.