Update deployment-config
It's not obvious what this is about, I've added a bit of text just to make this more readable. Its actually a crucial part of the system but I think its easy to overlook.

I think also that the callout boxes should just say Magento 1.x note or something similar so that people know to look in these boxes to see how it compares to 1.x. Having that comparison is a big help to devs, it feels like the callouts are too wordy at present.

This is just a suggestion.