Making deployment configuration more readable/relevant
This page really needs a re-write. I've done a few bits but I feel like it's written from a technical perspective and needs to be altered to be written from a user perspective.

For instance headings like 'config.php contents' mean nothing, a heading that says how to configure modules installed means a lot more.  I'm not sure that I care about knowing that '`config.php` and `env.php` are PHP files that return a multi-dimensional associative array' - is this relevant?  Feels like there is a lot of noise here and the true message about the importance of these files is being lost.