Reformat the Admin UI FAQ page
Instead of using 'Q' and 'A' to indicate questions and answers, I have
used the h3 headers to make the question stand out from the answers, which
seems to be a more common approach to FAQ pages.
The line breaks should be sufficient to distinguish question and
answer pairs.

I have also added a dynamic table of contents for the different questions
for easy navigation as the page fills up with more questions.