Change the Admin FAQ menu title and edit a question
I changed the Admin FAQ menu title to be Admin UI and UX FAQ to specify
the type of content that will be on that page.

A quoted question with a quoted answer was modified so that the question
is now more general and the quoted answer stands out more in a block quote.