Reformat and edit composer-integration page
I have replaced the table with a list of the different types of composer.json files
so that the content displays better. The location and name have been changed so that
they are less Magento focused and more general component development focused.

Sections have been moved around to improve flow and sequence of information.