Fix the search box bug
The cause of the bug is from the use of the searchResultsState cookie. When
a search is initiated, the cookie is enabled. There was logic in the code that
called the function to show the search results box when the cookie was set. So
when a search result was clicked and it opened up a new window, the cookie was
never reset and an empty search result box was displayed in the new window.

This bug only happened sporadically because the cookie was unique per section. For
example, if you did a search and you were in the architecture section, if you clicked
on a search result that was in the same section, you would see this bug. Otherwise,
you would not.

Changes in this commit involved removing the check for the cookie since it didn't
really do anything except cause the bug.

Code has been cleaned up some by moving initialization code under the document ready
function and moving the functions outside of it.