Apply CSS style rules to the collapsible elements
A collapsible.scss file was created and imported into app.scss

Gulp was used to minify the new styles to replace the old app.css.

Not sure what was changed in print.css. I suspect there were some scss rules
that were not minified and applied prior to checking out the code.