Add new page variable 'guide_version' alongside 'version'
The 'guide_version' will be dynamic and based on where the target html
will ber. This will be used to group all items in the same version.

The 'version' will be based on the front-matter value of the original .md
file. This will determine what the real source will be after following
symlinks.