Add tab functionality
Topics with the same tabgroup defined in the front matter will now
have tabbed navigation in the header area.

Introduces new front matter variable `tabgroup` to group together
pages for tabbed navigation.

Introduces new front matter variable `tablabel` to use as the tab
label for that page in the tabbed navigation. If this is not defined,
the page title will be used.

Introduces new front matter variable `tabweight` to use for ordering
tabs. Tabs are ordered least to greatest. If this variable is not
defined, the tab is considered to be weightless and have a value
of 0.

Squashed commit messages:
* Create tabs template snippet and apply to page header
* Style the tabs to look more tab-like
* Fix tab size
* Refactor tab implementation to use tabgroups and tablabels
* Add logic to use page title for label if tablabel is not defined
* Add sorting by tabweight functionality