Update css-practice.md
There is a lack of detail regarding the need to clear the static files cache....I spend hours following this specific section of the documentation (while in Developer mode) with *all* cache disabled, with zero effect...until I clicked on "Flush Static Files Cache". There should be clearer notification regarding this additional cache needing to be cleared at the beginning of all sections that related to css customization.