Fix a mistake in SSH port forwarding command
The OpenSSH official documentation says:

The -R option:
Specifies that connections to the given TCP port or Unix socket on the
remote (server) host are to be forwarded to the local side.

The -L option:
Specifies that connections to the given TCP port or Unix socket on the
local (client) host are to be forwarded to the given host and port, or
Unix socket, on the remote side.

The reference: https://man.openbsd.org/ssh

So, the -R option redirects a local port to a remote machine, and -L
redirects a remote port to a local machine. When we use Xdebug we need
the remote port to be forwarded to the local machine. That's why we need
the -L option instead of -R one.

I've checked other cloud articles and found that we use SSH port
forwarding only in the following article (except Xdebug):
guides/v2.0/cloud/project/project-conf-files_services-rabbit.md
and that article has correct "-L" flag.