Proposed partial cleanup to Layout overview-Introduction
I tightened up the first part of the introduction of the Layout overview section to introduce and focus on the key concepts: layout, containers, and blocks. I also updated the container and block images by consolidating them into one numbered image that illustrates the three concepts side by side and details those concepts by reference those numbers below the image. I also removed the MVC information from the beginning as it was a bit out of place and distracted from the explanation of the main topics.

Reorganization of the other topics within this overview is also needed, but not made in this commit.