Update README.md
The steps for building the devdocs locally on Linux (I'm going to say it's the same for Mac/Windows if using the corresponding "native Docker apps" ... https://docs.docker.com/docker-for-windows/ or https://docs.docker.com/docker-for-mac/ ), are outdated and incomplete.

For instance, the above README.md has references to a repository with "everything necessary to run Jekyll3" .. that's probably true, but it doesn't contain everything needed to get the DevDocs built locally ... and it certainly doesn't contain information on where to place this very repository in relation to the "docker-for-devdocs" repository ....

Bottom line, this very repository (github.com/magento/devdocs) contains a "Dockerfile" and "docker-compose.yml" file. Running the appropriate "docker-compose up" command *does* build everything needed to build a local copy of the devdocs. I just verified all of what I'm reporting is accurate, as I just went through the process. I have updated this file to include as many helpful instructions and links as possible to help with anyone's barrier to using Docker.