Update mage-dirs.md
The previous sample configuration doesn't work, as the custom array is overwritten by the default values. In order for this to work, the custom values need to come after the default system values ... and in the above case, since the example wasn't meant to overwrite the entire array, targeted array key/value pairs are required.