Update ms_nginx.md
This portion of the documentation is missing the actual touch point that passes the MAGE_RUN_TYPE and MAGE_RUN_CODE Nginx variables into php. I have updated this page to show a way to include these variables, as following the directions as they are currently presented does *not* successfully allow for someone to accomplish the task at hand.

I have also included $MAGE_RUN_TYPE in the Nginx vhost files to show how one could specify website vs store, if the desire was to pass in a store code vs a website code.