Remove wrong and irrelevant information
Magento 2 has never shipped using `Zend_Cache_Backend_File`.  
It uses Colin Molenhour's `Cm_Cache_Backend_File`. The whole first paragraph `Issues with Zend_Cache_Backend_File` is relevant to Magento 1 only, so ir should be removed

This belief the file system is slow a myth, and this page furthers it. It is based on the fact that the `Zend_Cache_Backend_File` implementation used in Magento 1 was crap. The file system is, in fact, a highly efficient key value store. It is cached in memory, too, so access is incredibly fast.  
In fact, on single machine instances `Cm_Cache_Backend_File` scales better than the redis cache backend `Cm_Cache_Backend_Redis` (also by Colin).
The overhead is much lower than using tcp over sockets to connect to a single threaded redis instance.
And if no socket connection is used to access redis but it goes over the actuall network sockets, it is much much much slower than the file system.