Added PHP Modifiers to UI Component request flow
In the list that details the UI Component request flow, I added in the step where PHP Modifier classes can modify the JSON before it is delivered to the browser. I'm assuming modifiers work *after* the XML is converted to JSON, based on the fact their modifyData() and modifyMeta() methods return JSON, but I could definitely be wrong.