Added description for system attribute option
The option reference for EAV attributes did not contain any information about the system option. A short description is added to the reference table.

Additionally a notice was added that system attributes prevent a custom attribute from being added to the `custom_attributes` array. This was not directly explained previously.