Fix custom logo example
If a `logo_src` argument is passed, this will directly use the string as
the image SRC, which is incorrect in this example (and probably
incorrect most of the time).

If a `logo_file` argument is passed, then Magento will correctly use a
fully qualified URL with static versioning, etc.