Updated inter server communication section
This is initially confusing because users would assume that they can skip this step if they had their database and apache on the same server. While this did allow the user to complete the installation process, it still prevented curl being used by any of the necessary integrations such as payment and shipping. This update will allow users to make a more accurate decision at this point.