Added link to custom cron reference
People may be directed to the documentation on how to use the cli subcommand 'cron', when they are using a search engine to find out how to create cron jobs. But on that page there is not explained about how to add custom cron jobs.

So I have added a link to the page where more information can be found.