Update BrainTree Statement
This is a contradictory statement to the following on this page itself. It is also not included in this blog post. URL: https://community.magento.com/t5/Magento-DevBlog/Deprecation-of-Magento-core-payment-integrations/ba-p/426445?_ga=2.145131826.704475700.1602302487-695246737.1590919082

This bundled extension replaces our core Braintree integration, which has been removed in this release. See Braintree Payments.
This is also on the same page