Change in correct path
I was trying to set up Varnish and was instructed to change the probe url from /pub/health_check.php to health_check.php. Unfortunately this did not work, I tried prepending a '/' which worked. So I think this should be changed in the docs before others waste time on this small thing. So it should be '/health_check.php' instead of 'health_check.php'