Update resources.md
Changing Magento version from v2.3 to v2.4 in links.
Added updated link for Slack channels, Which will redirect the user to the Slack Application instead of opening the Slack channel in the browser. If the user is not having a Slack Application, then the Slack channel will open in the browser.
Updated channel descriptions.
Added beginners channel and its description in recommended for new members section.