Update best-practices.md
Updated Magento version from v2.3 to v2.4
Added updated link for devdocs Slack channel, Which will redirect the user to the Slack Application instead of opening the Slack channel in the browser. If the user is not having a Slack Application, then devdocs Slack channel will open in the browser. The purpose is to gain more traffic on the Slack Channel.