Remove setup_version
`setup_version` is used for very specific types of `Setup/*.php` scripts. Not all modules make use of such scripts. Including reference to it here in this guide serves only as a distraction as there is no reference to it explaining what it does, nor should there as that is not the topic of this guide.